+++ sdk-okhttp-client/src/test/java/io/nem/symbol/sdk/infrastructure/okhttp/BlockRepositoryOkHttpImplTest.java
@@ -75,9 +75,9 @@ public class BlockRepositoryOkHttpImplTest extends AbstractOkHttpRespositoryTest
     BlockInfoDTO dto = new BlockInfoDTO();
     BlockMetaDTO metaDTO = new BlockMetaDTO();
     metaDTO.setHash("someHash");
-    metaDTO.setNumTransactions(10);
+    metaDTO.setTransactionsCount(10);
     metaDTO.setGenerationHash("generationHash");
-    metaDTO.setNumStatements(20);
+    metaDTO.setStatementsCount(20);
     metaDTO.setStateHashSubCacheMerkleRoots(Arrays.asList("string1", "string2"));
     metaDTO.setTotalFee(BigInteger.valueOf(8L));
 
@@ -113,9 +113,9 @@ public class BlockRepositoryOkHttpImplTest extends AbstractOkHttpRespositoryTest
     Assertions.assertEquals(NetworkType.MIJIN_TEST, info.getNetworkType());
     Assertions.assertEquals(BigInteger.valueOf(9L), info.getHeight());
     Assertions.assertEquals(metaDTO.getHash(), info.getHash());
-    Assertions.assertEquals(metaDTO.getNumTransactions(), info.getNumTransactions());
+    Assertions.assertEquals(metaDTO.getTransactionsCount(), info.getTransactionsCount());
     Assertions.assertEquals(metaDTO.getGenerationHash(), info.getGenerationHash());
-    Assertions.assertEquals(metaDTO.getNumStatements(), info.getNumStatements().get());
+    Assertions.assertEquals(metaDTO.getStatementsCount(), info.getStatementsCount());
     Assertions.assertEquals(
         metaDTO.getStateHashSubCacheMerkleRoots(), info.getSubCacheMerkleRoots());
     Assertions.assertEquals(metaDTO.getTotalFee(), info.getTotalFee());
@@ -132,9 +132,9 @@ public class BlockRepositoryOkHttpImplTest extends AbstractOkHttpRespositoryTest
     BlockInfoDTO dto = new BlockInfoDTO();
     BlockMetaDTO metaDTO = new BlockMetaDTO();
     metaDTO.setHash("someHash");
-    metaDTO.setNumTransactions(10);
+    metaDTO.setTransactionsCount(10);
     metaDTO.setGenerationHash("generationHash");
-    metaDTO.setNumStatements(20);
+    metaDTO.setStatementsCount(20);
     metaDTO.setStateHashSubCacheMerkleRoots(Arrays.asList("string1", "string2"));
     metaDTO.setTotalFee(BigInteger.valueOf(8));
 
@@ -176,9 +176,9 @@ public class BlockRepositoryOkHttpImplTest extends AbstractOkHttpRespositoryTest
     Assertions.assertEquals(NetworkType.MIJIN_TEST, info.getNetworkType());
     Assertions.assertEquals(BigInteger.valueOf(9L), info.getHeight());
     Assertions.assertEquals(metaDTO.getHash(), info.getHash());
-    Assertions.assertEquals(metaDTO.getNumTransactions(), info.getNumTransactions());
+    Assertions.assertEquals(metaDTO.getTransactionsCount(), info.getTransactionsCount());
     Assertions.assertEquals(metaDTO.getGenerationHash(), info.getGenerationHash());
-    Assertions.assertEquals(metaDTO.getNumTransactions(), info.getNumTransactions());
+    Assertions.assertEquals(metaDTO.getTransactionsCount(), info.getTransactionsCount());
     Assertions.assertEquals(
         metaDTO.getStateHashSubCacheMerkleRoots(), info.getSubCacheMerkleRoots());
     Assertions.assertEquals(metaDTO.getTotalFee(), info.getTotalFee());